refactor graphics-test and graphics capabilities query
