+++ .vscode/settings.json
@@ -32,6 +32,9 @@
         "limits": "c",
         "semaphore": "c",
         "algorithm": "c",
-        "cstring": "c"
+        "cstring": "c",
+        "graphics-test.h": "c",
+        "sysexits.h": "c",
+        "execinfo.h": "c"
     }
 }
