+++ build.zig
@@ -76,7 +76,7 @@ pub fn build(b: *std.Build) void {
     registerExample(b, target, optimize, staticLib, run, "events", b.path("examples/events.c"));
     registerExample(b, target, optimize, staticLib, run, "rawgl", b.path("examples/rawgl.c"));
     registerExample(b, target, optimize, staticLib, run, "graphics", b.path("examples/graphics.c"));
-    registerExample(b, target, optimize, staticLib, run, "graphicsTest", b.path("examples/graphics-test.c"));
+    registerExample(b, target, optimize, staticLib, run, "graphicsTest", b.path("examples/graphics-test/graphics-test.c"));
 }
 
 fn registerExample(b: *std.Build, target: std.Build.ResolvedTarget, optimize: std.builtin.OptimizeMode, pinclib: *std.Build.Step.Compile, run: bool, comptime name: []const u8, sourceFile: std.Build.LazyPath) void {
